pyproject.toml
[tool.poetry]
name = "web-scraping-ml"
version = "0.1.0"
description = ""
authors = ["Jonatan Gomez <andresgomez1796@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.3"
python-telegram-bot = "^21.10"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
